fixed the upper limit problem
because the sensor is loose because for some reason we still don't have a second working belt. And because of that when we dont press well on the cable the value jumps to 1023 which is much higher than the average high value form the sensor which around 400. So I had the idea of creating a new buffer that stores the last 200 arduino sensor values and then get their Max. This way if the value jumps to 1023 after a while it will use again the new max because 1023 will dequeued. But then there's another problem if the stay too long on a one value in an attempt to stay on .5 let's say. The max dropps and the range becomes smaller again. So I though of another idea. To lower the resistance in the circuit so the range value of the sensor is 1010 to 1023. So it's not a problem if the cable disconnect the difference is not so big compared to the normal upperlimit