new calculation for the tilt
2*value-1 (we get a signed value)
then we add left and right
so they could cancel each other out or add each.
pitch is relative to roll. Because if roll clockwise we should pitch up in the whale reference. same if we roll to the right same. so pitch=abs(roll). if value from both are equal the whale stops rolling. But then the pitch is value1 + value2. So if they are both positive pitch>0. if both <0. pitch>0